Fixing reservation bug

+++ python-project-starter-main/app/forms/new_venue_form.py
@@ -16,7 +16,7 @@ class NewVenueForm(FlaskForm):
     has_rooftop = BooleanField('has_rooftop')
     has_heated_outdoor_seating = BooleanField('has_heated_outdoor_seating')
     has_liquor = BooleanField('has_liquor')
-    price = FloatField('price')
+    price = IntegerField('price')
     
     created_at = DateTimeField('created_at')
     update_at = DateTimeField('update_at')
